payment processing

+++ pkg/models/user.model.go
@@ -52,13 +52,12 @@ type User struct {
 	Status                 string     `bson:"status,omitempty" json:"status"`
 	Program                string     `bson:"program,omitempty" json:"program"`
 	ProgramOption          string     `bson:"programOption,omitempty" json:"programOption"`
-	Wallet                 int        `bson:"wallet,omitempty" json:"wallet"`
+	Wallet                 float64    `bson:"wallet,omitempty" json:"wallet"`
 
 	TimeZone        *time.Time        `bson:"timeZone,omitempty" json:"timeZone,omitempty"`
 	AuthoredCourses CourseAuthorList  `bson:"authoredCourses,omitempty" json:"authoredCourses,omitempty"`
 	Courses         StudentCourseList `bson:"courses,omitempty" json:"courses,omitempty"`
 }
-
 
 //Bio is a model for Bios table
 type BioData struct {
@@ -87,11 +86,11 @@ type Qualification struct {
 
 // background is a model for backgrounds table
 type Background struct {
-	BornAgain        string   `bson:"bornAgain,omitempty" json:"bornAgain,omitempty"`
+	BornAgain        string `bson:"bornAgain,omitempty" json:"bornAgain,omitempty"`
 	SalvationBrief   string `bson:"briefSalvation,omitempty" json:"briefSalvation,omitempty"`
 	GodsWorkings     string `bson:"godsWorkings,omitempty" json:"godsWorkings"`
 	GodsCall         string `bson:"godsCall,omitempty" json:"godsCall"`
-	IntoMinistry     string   `bson:"intoMinistry,omitempty" json:"intoMinistry"`
+	IntoMinistry     string `bson:"intoMinistry,omitempty" json:"intoMinistry"`
 	SpiritualGifts   string `bson:"SpiritualGifts,omitempty" json:"spiritualGifts"`
 	Reason           string `bson:"reason,omitempty" json:"reason"`
 	ChurchName       string `bson:"churchName,omitempty" json:"churchName"`
@@ -100,24 +99,24 @@ type Background struct {
 	PastorEmail      string `bson:"pastorEmail,omitempty" json:"pastorEmail"`
 	PastorPhone      string `bson:"pastorPhone,omitempty" json:"pastorPhone"`
 	ChurchInvolved   string `bson:"churchInvolve,omitempty" json:"churchInvolve"`
-	WaterBaptism     string   `bson:"waterBaptism,omitempty" json:"waterBaptism"`
+	WaterBaptism     string `bson:"waterBaptism,omitempty" json:"waterBaptism"`
 	BaptismDate      string `bson:"baptismDate,omitempty" json:"baptismDate"`
-	HolyGhostBaptism string   `bson:"holyghostBaptism,omitempty" json:"holyghostBaptism"`
+	HolyGhostBaptism string `bson:"holyghostBaptism,omitempty" json:"holyghostBaptism"`
 }
 
 // Health is a model for Healths table
 type Health struct {
-	Disability             string   `bson:"disability,omitempty" json:"disability,omitempty"`
-	Nervousillness         string   `bson:"nervous,omitempty" json:"nervous,omitempty"`
-	Anorexia               string   `bson:"anorexic,omitempty" json:"anorexic,omitempty"`
-	Diabetese              string   `bson:"diabetic,omitempty" json:"diabetic,omitempty"`
-	Epilepsy               string   `bson:"epileptic,omitempty" json:"epileptic,omitempty"`
-	StomachUlcers          string   `bson:"stomachUlcer,omitempty" json:"stomachUlcer,omitempty"`
-	SpecialDiet            string   `bson:"specilaDiet,omitempty" json:"specialDiet,omitempty"`
-	LearningDisability     string   `bson:"learningDisability,omitempty" json:"learningDisability,omitempty"`
-	UsedIllDrug            string   `bson:"illegalDrugs,omitempty" json:"illegalDrug,omitempty"`
-	DrugAddiction          string   `bson:"drugAddiction,omitempty" json:"drugAddiction,omitempty"`
-	HadSurgery             string   `bson:"surgery,omitempty" json:"surgery,omitempty"`
+	Disability             string `bson:"disability,omitempty" json:"disability,omitempty"`
+	Nervousillness         string `bson:"nervous,omitempty" json:"nervous,omitempty"`
+	Anorexia               string `bson:"anorexic,omitempty" json:"anorexic,omitempty"`
+	Diabetese              string `bson:"diabetic,omitempty" json:"diabetic,omitempty"`
+	Epilepsy               string `bson:"epileptic,omitempty" json:"epileptic,omitempty"`
+	StomachUlcers          string `bson:"stomachUlcer,omitempty" json:"stomachUlcer,omitempty"`
+	SpecialDiet            string `bson:"specilaDiet,omitempty" json:"specialDiet,omitempty"`
+	LearningDisability     string `bson:"learningDisability,omitempty" json:"learningDisability,omitempty"`
+	UsedIllDrug            string `bson:"illegalDrugs,omitempty" json:"illegalDrug,omitempty"`
+	DrugAddiction          string `bson:"drugAddiction,omitempty" json:"drugAddiction,omitempty"`
+	HadSurgery             string `bson:"surgery,omitempty" json:"surgery,omitempty"`
 	HealthIssueDescription string `bson:"healthIssues,omitempty" json:"healthIssues"`
 }
